Separate out code into http.go file and complete implementation.
